Fixed case where orientation causes list overlaps in schedules menu. Added dialog to create new menu. Code to insert new menu in database still needs to be implemented.

+++ schedushare-android/SchedUShare/src/com/schedushare/android/MainMenuActivity.java
@@ -15,14 +15,16 @@ public class MainMenuActivity extends RoboActivity {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         
-        SchedulesDataSource dataSource = new SchedulesDataSource(this);
-        dataSource.open();
-        dataSource.dropAllTables();
-        dataSource.createUser(1, "oscarlee");
-        for (int i = 0; i < 100; i++) {
-        	dataSource.createSchedule(i, " schedule" + i + " ", true, 1, "whatever");
+        if (savedInstanceState == null) {
+	        SchedulesDataSource dataSource = new SchedulesDataSource(this);
+	        dataSource.open();
+	        dataSource.dropAllTables();
+	        dataSource.createUser(1, "oscarlee");
+	        for (int i = 0; i < 100; i++) {
+	        	dataSource.createSchedule(i, " schedule" + i + " ", true, 1, "whatever");
+	        }
+	        dataSource.close();
         }
-        dataSource.close();
     }
     
     // Called when user clicks schedules button.
